Renamed observation_bandwidth_hz in Correlator and Voltage contexts to bandwidth_hz to distinguish from metafits observation_bandwidth_hz. Fixed gps_time datatype and fixed more tests.

+++ src/voltage_files/error.rs
@@ -17,7 +17,7 @@ pub enum VoltageFileError {
     Mixture,
 
     #[error(r#"There are missing gps times- expected {expected} got {got}"#)]
-    GpsTimeMissing { expected: usize, got: usize },
+    GpsTimeMissing { expected: u64, got: u64 },
 
     #[error(r#"There are an uneven number of channel (files) across all of the gps times- expected {expected} got {got}"#)]
     UnevenChannelsForGpsTime { expected: u8, got: u8 },
